+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - `jung-esri`: translates Esri `drawingInfo` into Mapbox GL style layers plus a list of what could not be translated.
+- `jung-esri`: `esriPMS` and `esriPFS` symbols with inline `imageData` become `icon-image` and `fill-pattern` layers plus a map of named data uri images to register.
 
 ## [0.1.0] - 2026-05-30